Fix show only practiced tests checkbox

+++ LerenTypen/Database.cs
@@ -1558,7 +1558,7 @@ namespace LerenTypen
                                 while (reader.Read())
                                 {
                                     //add all the found data to a list
-                                    queryResult.Add(new TestTable(counter, reader.GetString(2), reader.GetInt32(4), reader.GetInt32(5), GetAmountOfWordsFromTest(reader.GetInt32(0)), reader.GetInt32(3), reader.GetString(6)));
+                                    queryResult.Add(new TestTable(counter, reader.GetString(2), reader.GetInt32(4), reader.GetInt32(5), GetAmountOfWordsFromTest(reader.GetInt32(0)), reader.GetInt16(3), reader.GetString(6)));
                                     counter++;
                                 }
                                 reader.NextResult();
